fix(label): update label tests
