+++ packages/paste-core/components/label/package.json
@@ -2,7 +2,7 @@
   "name": "@twilio-paste/label",
   "version": "1.0.0",
   "category": "user input",
-  "status": "beta",
+  "status": "production",
   "description": "Label used to name Paste form fields",
   "author": "Twilio Inc.",
   "license": "MIT",
