+++ packages/paste-core/components/label/src/Label.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import {styled, css} from '@twilio-paste/styling-library';
 import {Box, safelySpreadBoxProps} from '@twilio-paste/box';
 import {Flex} from '@twilio-paste/flex';
 import {Text} from '@twilio-paste/text';
 import {ScreenReaderOnly} from '@twilio-paste/screen-reader-only';
 import {TextColor} from '@twilio-paste/style-props';
 
+export type LabelVariants = 'default' | 'inverse';
 export interface LabelProps extends React.LabelHTMLAttributes<HTMLLabelElement> {
   as?: 'label' | 'legend';
   children: NonNullable<React.ReactNode>;
@@ -14,21 +14,27 @@ export interface LabelProps extends React.LabelHTMLAttributes<HTMLLabelElement> 
   htmlFor: string | undefined;
   marginBottom?: 'space0';
   required?: boolean;
-  variant?: 'default' | 'inverse';
+  variant?: LabelVariants;
 }
 
-// eslint-disable-next-line emotion/syntax-preference
-const StyledRequiredDot = styled(Text)(
-  css({
-    backgroundColor: 'colorBackgroundRequired',
-    borderRadius: '50%',
-    cursor: 'pointer',
-    display: 'block',
-    height: '4px',
-    marginRight: 'space20',
-    width: '4px',
-  })
-);
+export const RequiredDot: React.FC = props => {
+  return (
+    <Box
+      {...props}
+      as="span"
+      backgroundColor="colorBackgroundRequired"
+      borderRadius="borderRadiusCircle"
+      cursor="pointer"
+      display="block"
+      height="4px"
+      lineHeight="lineHeight30"
+      marginRight="space20"
+      width="4px"
+    >
+      <ScreenReaderOnly>Required: </ScreenReaderOnly>
+    </Box>
+  );
+};
 
 const Label: React.FC<LabelProps> = ({as, marginBottom, required, disabled, children, variant, ...props}) => {
   let textColor = 'colorText' as TextColor;
@@ -53,11 +59,7 @@ const Label: React.FC<LabelProps> = ({as, marginBottom, required, disabled, chil
       textTransform="none"
     >
       <Flex as="span" vAlignContent="center">
-        {required ? (
-          <StyledRequiredDot as="span" lineHeight="lineHeight30">
-            <ScreenReaderOnly>Required: </ScreenReaderOnly>
-          </StyledRequiredDot>
-        ) : null}
+        {required ? <RequiredDot /> : null}
         <Text
           as="span"
           fontSize="fontSize30"
