+++ packages/paste-core/components/label/tsconfig.build.json
@@ -18,9 +18,6 @@
       "path": "../../layout/flex"
     },
     {
-      "path": "../../../paste-icons"
-    },
-    {
       "path": "../screen-reader-only"
     },
     {
